requirements.txt
streamlit
requests
opencv-python
Pillow
numpy
onnxruntime
ffmpeg-python